Add files via upload
From the survey conducted, obtained data with a number of interest and demographic information of a region. The data has 425,332 rows and 9 columns. The columns are id, name, age, age_range, province, city, gender, interest and interest_detail.

Of the 425,332 data, there are a number of duplicate data and missing values. The client requests that all survey data for gender information be filled in completely.

Based on these problems, I used several machine learning methods (NBC, Random Forest, KNN) and used the name column as a feature to solve this problem.

The steps taken are Preprocessing/Cleansing dataset, Training Model, Evaluation, and Predict.

The best accuracy is 82.9%.